button: replace css state with react state
also split files for clarity
