+++ package.json
@@ -48,6 +48,7 @@
     "react-dom": "^16.3.1",
     "react-emotion": "^9.1.1",
     "react-node-resolver": "^1.0.1",
+    "react-pseudo-state": "^1.0.1",
     "react-radios": "^1.1.0",
     "react-router-dom": "^4.2.2",
     "react-scrolllock": "^2.0.2",
