+++ packages/ui/src/primitives/buttons/index.js
@@ -1,0 +1,2 @@
+export { LoadingButton } from './loading';
+export { Button } from './primitives';
