+++ packages/ui/src/primitives/buttons/loading.js
@@ -1,0 +1,60 @@
+// @flow
+
+import React from 'react';
+import styled from 'react-emotion';
+
+import { LoadingIndicator, LoadingSpinner } from '../loading';
+import { Button, type ButtonProps } from './primitives';
+
+// Styled
+
+const LoadingButtonWrapper = styled(Button)({ position: 'relative' });
+const LoadingIndicatorWrapper = styled.div({
+  left: '50%',
+  position: 'absolute',
+  top: '50%',
+  transform: 'translate(-50%, -50%)',
+});
+
+function getAppearance(appearance) {
+  if (appearance === 'default') return 'dark';
+  return 'inverted';
+}
+
+// Export
+
+type Loading = ButtonProps & {
+  isLoading: boolean,
+  indicatorVariant: 'spinner' | 'dots',
+};
+export const LoadingButton = ({
+  children,
+  indicatorVariant,
+  isLoading,
+  ...props
+}: Loading) => {
+  const appearance = getAppearance(props.appearance);
+  const textCSS = isLoading ? { visibility: 'hidden' } : null;
+  const isSpinner = indicatorVariant === 'spinner';
+
+  return (
+    <LoadingButtonWrapper {...props}>
+      {isLoading ? (
+        <LoadingIndicatorWrapper>
+          {isSpinner ? (
+            <LoadingSpinner appearance={appearance} size={16} />
+          ) : (
+            <LoadingIndicator appearance={appearance} size={4} />
+          )}
+        </LoadingIndicatorWrapper>
+      ) : null}
+      <span css={textCSS}>{children}</span>
+    </LoadingButtonWrapper>
+  );
+};
+LoadingButton.defaultProps = {
+  appearance: 'default',
+  isLoading: false,
+  variant: 'bold',
+  indicatorVariant: 'dots',
+};
